fix: ga 관련 코드 수정
